requirements.txt
streamlit
pandas
requests
python-dotenv
sqlparse
mysql-connector-python
SQLAlchemy
PyMySQL